+++ server/src/model/picture_model.js
@@ -1,4 +1,4 @@
-const mysql = require("../database/connect.js");
+const mysql = require("../../database/connect.js");
 
 const Picture = function (picture) {
     this.id = picture.id;
