+++ server/src/model/user_model.js
@@ -1,4 +1,4 @@
-const mysql = require("../database/connect.js");
+const mysql = require("../../database/connect.js");
 const table = "USER";
 const passport = require('passport');
